Ending index page development
