+++ src/sass/style.scss
@@ -276,12 +276,17 @@
   }
 }
 
+footer .footer {
+  background: linear-gradient(90deg, #213f8c 50%, #1e3980 75%, #1b3373 100%);
+}
+
 .transition {
   transition: 0.3s all;
 }
 
 *::-webkit-scrollbar {
   width: 10px;
+  height: 0;
 }
 *::-webkit-scrollbar-thumb {
   background-color: $awe-blue-third;
