+++ src/sass/utilities/_box-shadow.scss
@@ -1,16 +1,19 @@
 @each $prop, $abbrev in (box-shadow: bx) {
   @each $key, $value in $theme-colors {
     .#{$abbrev}-#{$key}-left-bottom {
+      --box-shadow-color: #{$value};
       #{$prop}: -8px 8px $value !important;
     }
     .#{$abbrev}-#{$key}-right-top {
+      --box-shadow-color: #{$value};
       #{$prop}: 8px -8px $value !important;
     }
-
     .#{$abbrev}-#{$key}-left-top {
+      --box-shadow-color: #{$value};
       #{$prop}: -8px -8px $value !important;
     }
     .#{$abbrev}-#{$key}-right-bottom {
+      --box-shadow-color: #{$value};
       #{$prop}: 8px 8px $value !important;
     }
   }
@@ -49,6 +52,6 @@
   transition: 0.3s all !important;
 
   &:hover {
-    box-shadow: 0 0 transparent !important;
+    box-shadow: 0 0 var(--box-shadow-color) !important;
   }
 }
